+++ babel-register.js
