+++ package.json
@@ -4,7 +4,7 @@
   "description": "S3 Compatible Cloud Storage client",
   "source": "./src/main/minio.js",
   "main": "./dist/main/minio.js",
-  "types": "./dist/main/minio.d.ts",
+  "types": "./types/minio.d.ts",
   "scripts": {
     "format": "prettier -w .",
     "format-check": "prettier --list-different .",
@@ -13,7 +13,7 @@
     "lint": "eslint --ext js,mjs,cjs,ts ./",
     "lint-fix": "eslint --ext js,mjs,cjs,ts ./ --fix",
     "prepublishOnly": "npm test",
-    "functional": "mocha src/test/functional/functional-tests.js",
+    "functional": "mocha tests/functional/functional-tests.js",
     "type-check": "tsc --noEmit --emitDeclarationOnly false",
     "tsc": "tsc",
     "lint-staged": "lint-staged",
@@ -23,6 +23,7 @@
     "package.json",
     "./dist/",
     "./src/",
+    "./types/",
     "LICENSE",
     "README.md",
     "README_zh_CN.md",
@@ -79,6 +80,7 @@
     "xml2js": "^0.5.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.12.10",
     "@babel/core": "^7.12.10",
     "@babel/preset-env": "^7.12.10",
     "@babel/preset-typescript": "^7.21.4",
@@ -94,7 +96,6 @@
     "@typescript-eslint/eslint-plugin": "^5.57.1",
     "@typescript-eslint/parser": "^5.57.1",
     "@upleveled/babel-plugin-remove-node-prefix": "^1.0.4",
-    "babel": "^6.23.0",
     "babel-plugin-replace-import-extension": "^1.1.3",
     "babel-plugin-transform-replace-expressions": "^0.2.0",
     "chai": "^4.2.0",
