+++ src/AssumeRoleProvider.js
