+++ src/minio.js
