+++ src/object-uploader.js
