+++ src/signing.js
