+++ tests/functional/functional-tests.js
@@ -30,7 +30,7 @@ import splitFile from 'split-file'
 import superagent from 'superagent'
 import * as uuid from 'uuid'
 
-import { AssumeRoleProvider } from '../../main/AssumeRoleProvider.js'
+import { AssumeRoleProvider } from '../../src/AssumeRoleProvider.js'
 import {
   CopyDestinationOptions,
   CopySourceOptions,
@@ -38,8 +38,8 @@ import {
   getVersionId,
   isArray,
   removeDirAndFiles,
-} from '../../main/helpers.js'
-import * as minio from '../../main/minio.js'
+} from '../../src/helpers.js'
+import * as minio from '../../src/minio.js'
 
 const assert = chai.assert
