+++ tsconfig.json
@@ -14,7 +14,7 @@
     "emitDeclarationOnly": true,
     "sourceMap": true,
     "pretty": true,
-    "outDir": "./dist/main/"
-  },
-  "include": ["./src/main/*"]
+    "outDir": "./dist/main/",
+    "sourceRoot": "./src/"
+  }
 }
