+++ types/minio.d.ts
